Utilized Tone() for BEEGEES Melody
I utilized the built-in Arduino Tone() function and "pitches.h" header
to recreate the BeeGees melody. It looks a lot cleaner in the code, and
it's easier to figure out the melody (which is also now in a nice
looking array). The only down side is that the volume is about half. I
would love to hack the arduino Tone() function to be able to play the
tone on both sides of the buzzer. But for now, I think the cleaner code
is more important.