Lots of "trampoline" stuff
Whew, there is a lot here. I will do my best to remember. This work was
done a while ago.

Midi stuff. It now uses software serial lines to command a musical
instrument shield and trigger sounds when you jump. This makes it a bit
more interesting than what you can do with only the buzzer on board. I
also created some banks of sounds that go well together, and each time
you start a new round, it randomly chooses a bank (aka group of 4).

Triggering stuff - involves counter variables to keep track of "hang
time" and prevent false multi jumps. Also, this enables the triggering
to only occur when a trampoline has experienced hang time, and then
crossing the threshold "down" causes a trigger.

Includes Dunk tank stuff. This was a way to connect a single GPIO pin A3
to another system, and upon the pin going low for a period of time it
would enable the dunk tank. Used at AVC back in the day to dunk Nate and
plenty of other willing Funions.

Includes some smarts to avoid repeating a ton of the same trampoline
(aka buttons) in a sequence. This ensures each participant (if you are
playing one-player-per-trampoline) gets to have a couple jumps and isn't
just standing around waiting.

Added playback speed variable, to enable the playback to speed up as the
pattern gets longer. This is nice when you are at an event and the
players need to get the hang of it before it goes really fast.

That's about all I can see from reading through all the differences (and
remember). There is probably some more, but unfortunately, I was not in
the habit of github and making lots of commits at the time.

Here are my notes from at the top of the code. These will probably get
removed eventually, but are relevant at this commit.

Happy Jumping!

3/2/2016
Pete Lewis
Modifying code to work better for trampolines

1)
Fix bug on startup when all four players are standing on trampolines
-I think this is something to do with the variable keeping track of if
trampolines have been "up" meaning someone jumped up or steped off
completely.
- This is necessary so that you can stand on the trampolines or "walk"
from one to another. It makes the triggering occur only when a
trampoline crosses the threshold
- of high to low. FIXED 3/2/2016

2)
Increase playback speed as sequence grows. This will allow for faster
playing. We are planning on having groups of 4 play for 2 minutes, so
they need to be able to try a few times,
and potentially win a round within that time frame. ADDED 3/2/2016

3)
One of the tablah sounds is too low for the computer speakers to
actually make, so I need to change this to a sound that will be heard.
Also, it could be good to choose some more groups to add variety. FIXED
3/2/2016

4)
Add some code to avoid repeating the same button more than 2 times in a
sequence DONE 3/2/2016 - mostly happens in add_to_moves()

5)
Ensure that each button gets chosen an equal amount. This will ensure
that when you are playing four player trampolines, everyone gets to
jump.*

*I'm sure this is done, but not sure when it happened.